requirements.txt
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.4
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.1.8
cssselect==1.3.0
distro==1.9.0
exceptiongroup==1.3.0
faiss-cpu==1.11.0
fastapi==0.115.14
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.18.0
fsspec==2025.5.1
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.2
idna==3.10
jieba3k==0.35.1
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
lxml==6.0.0
lxml_html_clean==0.4.2
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.2.1
newspaper3k==0.2.8
nltk==3.9.1
numpy==2.0.2
openai==1.93.0
packaging==25.0
pillow==11.3.0
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
requests-file==2.1.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.13.1
sentence-transformers==5.0.0
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
starlette==0.46.2
sympy==1.14.0
threadpoolctl==3.6.0
tinysegmenter==0.3
tldextract==5.3.0
tokenizers==0.21.2
torch==2.7.1
tqdm==4.67.1
transformers==4.53.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
